+++ auth-db/queries/email-auth/index.ts
@@ -1,3 +1,4 @@
 export * from "./create.ts";
 export * from "./get-by-email.ts";
 export * from "./update-verification.ts";
+export * from "./update-forgot-password-token.ts";
